fix(runtime)!: match FFI widths between Zan externs and the C runtime (A2-0)
An audit of all 864 stdlib externs against the C definitions (_scratch is not
committed; the checker walks stdlib/*.zan and src/**/*.c) found 21 of our own
shims whose declared width no longer matches now that Zan `int` is i32. Each is
fixed on the side that carries the real semantics:

- status/boolean results stay 32-bit and the C shim says so: zan_io_socket_ready,
  zan_io_socket_alive, zan_io_connect_status, zan_dispatch_post, zan_thread_start
  and the shared-table set/delete/exists/match/destroy family return int32_t;
  zan_shared_table_create takes int32_t capacity/key_size
- genuinely 64-bit values become `long` in Zan: zan_io_socket_send/recv byte
  counts and lengths, and zan_monotonic_us

Stopwatch.GetMicroseconds() overflowed an i32 after 35 minutes of uptime and
GetMilliseconds() truncated GetTickCount64() after 24.8 days; both return `long`
now, as does the monotonic clock Socket uses for connect deadlines.
Socket.Send/Recv keep the C# `int` signature and narrow explicitly.

+++ src/runtime/rt_io.c
@@ -308,7 +308,7 @@ int64_t zan_io_socket_recv(intptr_t fd, void *buf, int64_t len,
 #endif
 }
 
-int64_t zan_io_socket_ready(intptr_t fd, int32_t write_ready) {
+int32_t zan_io_socket_ready(intptr_t fd, int32_t write_ready) {
     fd_set fds;
     FD_ZERO(&fds);
     struct timeval timeout = {0, 0};
@@ -328,7 +328,7 @@ int64_t zan_io_socket_ready(intptr_t fd, int32_t write_ready) {
 }
 
 /* See rt_io.h: whether `fd` is still an open socket. */
-int64_t zan_io_socket_alive(intptr_t fd) {
+int32_t zan_io_socket_alive(intptr_t fd) {
 #if defined(_WIN32)
     if ((SOCKET)fd == INVALID_SOCKET) return 0;
     int type = 0;
@@ -351,7 +351,7 @@ int64_t zan_io_socket_alive(intptr_t fd) {
  * populated until the exception set has actually been polled, so a bare
  * getsockopt(SO_ERROR) after the reactor wake reads 0 for a refused
  * connection. */
-int64_t zan_io_connect_status(intptr_t fd) {
+int32_t zan_io_connect_status(intptr_t fd) {
     fd_set wfds, efds;
     FD_ZERO(&wfds);
     FD_ZERO(&efds);
